Use the cli package for the CLI
We're now using the package https://github.com/urfave/cli to construct
the CLI. This has a number of advantages over hand-construction using
the "flags" package in the Go standard library. This dependency is
managed using the Glide package manager https://glide.sh/

This also include updates the README with information on how to use and
how to build getignore.