Big refactor, program in non-working state
Tries to make the HTTP requests happen in parallel, by using a buffered
channel, and issuing each request as a separate goroutine.

Also tries to simplify the output by taking a slice of strings, instead
of a channel. Adds a function to take the channel of request responses
and push them onto a slice.