Internalize channels into the HTTPGetter.
Also fixes the deadlock that occurs when the number of requested files
exceeds the maximum number of connections available (Issue #3).

At the moment, contents may be written to the output file out of order.
A later commit will add back the guarantee of contents being written in
order requested in the input.

As a side comment, I like this new interface for the getters — that they
accept a slice of desired file names and returns the contents and an
error, if any happened.